requirements.txt
requests>=2.28.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
python-dateutil>=2.8.2
tqdm>=4.65.0
plotly>=5.0.0

# ML + notebook execution/runtime deps
scikit-learn>=1.2.2
joblib>=1.2.0
nbconvert>=7.2.0
ipykernel>=6.22.0
